Remove unnecessary info and fix syslog-ng config
syslog-ng config now filters new messages with new prefix that is
default logging prefix for OpenWrt's firewall3. This means that now
logging is just question of enabling logging on given zone.

Example of uci configuration:
    config zone
	option name 'wan'
	list network 'wan'
	option input 'REJECT'
	option output 'ACCEPT'
	option forward 'REJECT'
	option log_limit '500/sec'
	option log '1'

The same change has to be done in syslog parser. New message format is:
REJECT wan in: IN=eth2 OUT= SRC=172.20.6.10 DST=172.20.6.255

This also removes some of the fields we do not use on server.
It is questionable if we need exact time of event so we do not send it.
Rule ID is no longer relevant as we do not match rules anymore.
Packet count was removed as well with simple package merging.
Flags are removed as they are not used on server at all.