Drop shell script to run nikola
This instead now runs python directly.

All arguments set in script are default values in argument parser. This
means that script was just duplicating them.

This also changes default behavior to not use random sleep. The option
-n and --now are preserved but effectivelly only --random-sleep has some
effect and that is enablement of random sleep. This means that cron now
uses that argument.