syslog_parser: drop directions and send only input connection
This was heavily discussed but we did not found out any usecase of
logging blocked traffice leaving router. In most cases there are going
to be no blocking rules for outgoing traffic so there are going to be no
logs for it as well. It is also questionable what exactly we would see
from such data.